refactor: move combobox into review specific module for profs
