--- src/common/components/Combobox/index.ts
@@ -1,2 +1,0 @@
-export * from "./Combobox";
-export * from "./Combobox.theme";
